Make building of tests optional
When ngtemplate is used as a build dependency it is useful to disable
building tests to speed up the build process.  Set the NGT_BUILD_TESTS
option to control this.  By default it is ON (i.e. tests are built) so
that the previous default is maintainted.

Setting the option may be done from a CMakeLists.txt file that lists
ngtemplate as a dependency:

	SET(NGT_BUILD_TESTS OFF)

It may also be set on the command line of cmake

	cmake -DNGT_BUILD_TESTS=OFF ...